# feat: pre-sampling