This PR only covers the single-machine training script. Pre-sampling improves the training throughput on GDELT on 4xA100 GPUs by up to 1.47x. 

Here are the complete results along with time breakdowns. The default hyperparameters, such as batch size and fanouts, were used. The pre-sampling time was computed by subtracting the total time from the time taken by other stages.


| TGN's cumulative time  | w/o pre-sampling  | w/ pre-sampling | w/ pre-sampling w/ GIL release| 
| --- | --- | --- | --- |
| Sample (second) | 107.81 | 146.75 | 73.21  | 
| Feature Fetching (second) |40.05 | 51.15  | 93.90|
| Memory Fetching (second) | 55.89 | 74.18 | 64.94 | 
| Model Train (second) | 133.27 | 139.96| 140.34 | 
| Memory Update (second) |29.52 | 29.24 | 28.46 |
| Throughput (samples/s) | 978710.12  | 870026.74 | 957765.28 (0.98x) | 


| TGAT's cumulative time  | w/o pre-sampling  | w/ pre-sampling | w/ pre-sampling w/ GIL release| 
| --- | --- | --- | --- |
| Sample (second) | 1615.54 | 1789.69 | 601.47 | 
| Feature Fetching (second)|1233.05 | 1176.75 |  1255.53 |
| Model Train (second)|  1139.23 | 1149.03 |  1132.83| 
| Throughput (samples/s )| 94681.67 | 93595.10  | 128832.98 (1.36x) |


| DySAT's cumulative time  | w/o pre-sampling  | w/ pre-sampling w/ GIL release| 
| --- | --- | --- | 
| Sample (second) | 1931.05 | 716.06 | 
| Feature Fetching (second)|688.48 | 787.35 |  
| Model Train (second)|  1942.29 | 2430.12 |  
| Throughput (samples/s )| 82520.29 | 97924.22 (1.19x) | 


| GraphSAGE's cumulative time  | w/o pre-sampling  | w/ pre-sampling w/ GIL release| 
| --- | --- | --- | 
| Sample (second) | 1966.26 | 495.08 | 
| Feature Fetching (second)|1539.45 | 1837.84  | 
| Model Train (second)|  941.65 | 298.84 | 
| Throughput (samples/s )| 99470.85 | 146083.56 (1.47x) | 


| GAT's cumulative time  | w/o pre-sampling  | w/ pre-sampling w/ GIL release| 
| --- | --- | --- | 
| Sample (second) | 1565.81 | 568.68 | 
| Feature Fetching (second)|1219.04 | 1434.58  | 
| Model Train (second)|  1139.23 | 967.24 |  
| Throughput (samples/s )| 101284.63 | 129671.03 (1.28x) |